+++ tests/Jackalope/Transport/DoctrineDBAL/ClientTest.php
@@ -236,6 +236,34 @@ class ClientTest extends FunctionalTestCase
         $this->addToAssertionCount(1);
     }
 
+    public function testDeleteProperties(): void
+    {
+        $root = $this->session->getNode('/');
+        $node = $root->addNode('delete-properties');
+        for ($i = 0; $i <= 1000; ++$i) {
+            $node->setProperty('property-'.$i, 'value-'.$i);
+        }
+
+        $this->session->save();
+        $this->assertSame(1002, \count($node->getProperties()));
+
+        for ($i = 501; $i <= 1000; ++$i) {
+            $node->setProperty('property-'.$i, null);
+        }
+
+        $this->session->save();
+        $this->session->refresh(false);
+        $node = $this->session->getNode('/delete-properties');
+
+        for ($i = 0; $i <= 1000; ++$i) {
+            $this->assertSame(
+                $i < 501,
+                $node->hasProperty('property-'.$i),
+                'Unexpected result for property "property-'.$i.'"'
+            );
+        }
+    }
+
     public function testPropertyLengthAttribute(): void
     {
         $rootNode = $this->session->getRootNode();