+++ CHANGELOG.md
@@ -21,6 +21,16 @@ Changelog
 1.x
 ===
 
+1.11.1
+------
+
+* Fix regression of 1.11.0: Reference deletion should work regardless of upper or lowercase of the type names in XML data.
+
+1.11.0
+------
+
+* Improve delete properties performance by replace DOMDocument with `xml_parse`.
+
 1.10.1
 ------